move test as separate module
